+++ euhedral-data-structures/src/test/java/euhedral/queues/PartitionedUnboundedMpscArrayQueueTest.java
@@ -70,7 +70,7 @@ class PartitionedUnboundedMpscArrayQueueTest {
 
             long deadline = System.nanoTime() + TimeUnit.SECONDS.toNanos(5);
             while (drained.sum() < batch * 8 && System.nanoTime() < deadline) {
-                int count = q.drain(consumer, 4096);
+                long count = q.drain(consumer, 4096);
                 drained.add(count);
                 Thread.yield();
             }
